moved objectid validation and lean to mongoose lecture, finished subdocs prep

+++ module10-mongodb/eCommerce-kick-off/controllers/ducks.js
@@ -7,7 +7,10 @@ const getAllDucks = async (req, res) => {
 };
 
 const createDuck = async (req, res) => {
-  const newDuck = await Duck.create(req.sanitizedBody);
+  const { owner, name, imgUrl, quote } = req.sanitizedBody;
+  //   if (!name || !imgUrl) throw new Error('Missing required fields', { cause: 400 });
+
+  const newDuck = await Duck.create({ owner, name, imgUrl, quote });
   res.status(201).json(newDuck);
 };
 
@@ -24,13 +27,15 @@ const getDuckById = async (req, res) => {
 };
 
 const updateDuck = async (req, res) => {
+  // const { userId } = req;
   const { id } = req.params;
-
   if (!isValidObjectId(id)) throw new Error('Invalid id', { cause: 400 });
 
   const duck = await Duck.findByIdAndUpdate(id, req.sanitizedBody, { new: true });
 
   if (!duck) throw new Error('Duck not found', { cause: 404 });
+
+  // if (userId !== duck.user.id) throw new Error('You are not authorized to update this duck', { cause: 403 });
 
   const duckWithOwner = await duck.populate('owner', 'firstName lastName');
   res.json(duckWithOwner);
@@ -38,6 +43,8 @@ const updateDuck = async (req, res) => {
 
 const deleteDuck = async (req, res) => {
   const { id } = req.params;
+
+  if (!isValidObjectId(id)) throw new Error('Invalid id', { cause: 400 });
 
   const duck = await Duck.findByIdAndDelete(id);
