+++ module10-mongodb/mongoose-lecture/controllers/users.js
@@ -1,7 +1,8 @@
+import { isValidObjectId } from 'mongoose';
 import User from '../models/User.js';
 
 export const getUsers = async (req, res) => {
-  const users = await User.find();
+  const users = await User.find().lean();
   res.json(users);
 };
 
@@ -23,7 +24,8 @@ export const getUserById = async (req, res) => {
   const {
     params: { id }
   } = req;
-  const user = await User.findById(id);
+  const user = await User.findById(id).lean();
+  if (!isValidObjectId(id)) throw new Error('Invalid id', { cause: 400 });
   if (!user) throw new Error('User not found', { cause: 404 });
   res.json(user);
 };
@@ -33,7 +35,7 @@ export const updateUser = async (req, res) => {
     sanitizedBody: { firstName, lastName, email },
     params: { id }
   } = req;
-
+  if (!isValidObjectId(id)) throw new Error('Invalid id', { cause: 400 });
   const user = await User.findByIdAndUpdate(id, req.sanitizedBody, { new: true });
   if (!user) throw new Error('User not found', { cause: 404 });
 
@@ -44,6 +46,7 @@ export const deleteUser = async (req, res) => {
   const {
     params: { id }
   } = req;
+  if (!isValidObjectId(id)) throw new Error('Invalid id', { cause: 400 });
   const user = await User.findByIdAndDelete(id);
 
   if (!user) throw new Error('User not found', { cause: 404 });
